LayoutTests/http/tests/referrer-policy-img/unsafe-url/cross-origin-http-UpgradeMixedContent.https-expected.txt

CONSOLE MESSAGE: The page at https://127.0.0.1:8443/referrer-policy-img/unsafe-url/cross-origin-http-UpgradeMixedContent.https.html requested insecure content from http://localhost:8000/referrer-policy/resources/image.py?expected=https://127.0.0.1:8443/referrer-policy-img/unsafe-url/cross-origin-http-UpgradeMixedContent.https.html. This content was automatically upgraded and should be served over HTTPS.

Tests the behavior of unsafe-url referrer policy when cross origin from HTTPS to HTTP.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS referrer is "https://127.0.0.1:8443/referrer-policy-img/unsafe-url/cross-origin-http-UpgradeMixedContent.https.html"
PASS successfullyParsed is true

TEST COMPLETE
